+++ .rubocop_todo.yml
@@ -155,6 +155,7 @@ Rails/NotNullColumn:
 # Offense count: 8
 Rails/OutputSafety:
   Exclude:
+    - 'app/controllers/sessions_controller.rb'
     - 'app/controllers/users_controller.rb'
     - 'app/helpers/application_helper.rb'
     - 'lib/rich_text.rb'
