+++ app/controllers/users_controller.rb
@@ -1,8 +1,10 @@
 class UsersController < ApplicationController
+  include SessionMethods
+
   layout "site"
 
   skip_before_action :verify_authenticity_token, :only => [:auth_success]
-  before_action :disable_terms_redirect, :only => [:terms, :save, :logout]
+  before_action :disable_terms_redirect, :only => [:terms, :save]
   before_action :authorize_web
   before_action :set_locale
   before_action :check_database_readable
@@ -11,7 +13,7 @@ class UsersController < ApplicationController
 
   before_action :require_self, :only => [:account]
   before_action :check_database_writable, :only => [:new, :account, :confirm, :confirm_email, :lost_password, :reset_password, :go_public]
-  before_action :require_cookies, :only => [:new, :login, :confirm]
+  before_action :require_cookies, :only => [:new, :confirm]
   before_action :lookup_user_by_name, :only => [:set_status, :destroy]
   before_action :allow_thirdparty_images, :only => [:show, :account]
 
@@ -269,38 +271,6 @@ class UsersController < ApplicationController
     end
   end
 
-  def login
-    append_content_security_policy_directives(
-      :form_action => %w[*]
-    )
-
-    session[:referer] = safe_referer(params[:referer]) if params[:referer]
-
-    if request.post?
-      session[:remember_me] ||= params[:remember_me]
-      password_authentication(params[:username], params[:password])
-    end
-  end
-
-  def logout
-    @title = t "users.logout.title"
-
-    if request.post?
-      if session[:token]
-        token = UserToken.find_by(:token => session[:token])
-        token&.destroy
-        session.delete(:token)
-      end
-      session.delete(:user)
-      session_expires_automatically
-      if params[:referer]
-        redirect_to safe_referer(params[:referer])
-      else
-        redirect_to :controller => "site", :action => "index"
-      end
-    end
-  end
-
   def confirm
     if request.post?
       token = UserToken.find_by(:token => params[:confirm_string])
@@ -515,93 +485,6 @@ class UsersController < ApplicationController
   private
 
   ##
-  # handle password authentication
-  def password_authentication(username, password)
-    if user = User.authenticate(:username => username, :password => password)
-      successful_login(user)
-    elsif user = User.authenticate(:username => username, :password => password, :pending => true)
-      unconfirmed_login(user)
-    elsif User.authenticate(:username => username, :password => password, :suspended => true)
-      failed_login t("users.login.account is suspended", :webmaster => "mailto:#{Settings.support_email}").html_safe, username
-    else
-      failed_login t("users.login.auth failure"), username
-    end
-  end
-
-  ##
-  # return the URL to use for authentication
-  def auth_url(provider, uid, referer = nil)
-    params = { :provider => provider }
-
-    params[:openid_url] = openid_expand_url(uid) if provider == "openid"
-
-    if referer.nil?
-      params[:origin] = request.path
-    else
-      params[:origin] = "#{request.path}?referer=#{CGI.escape(referer)}"
-      params[:referer] = referer
-    end
-
-    auth_path(params)
-  end
-
-  ##
-  # special case some common OpenID providers by applying heuristics to
-  # try and come up with the correct URL based on what the user entered
-  def openid_expand_url(openid_url)
-    if openid_url.nil?
-      nil
-    elsif openid_url.match(%r{(.*)gmail.com(/?)$}) || openid_url.match(%r{(.*)googlemail.com(/?)$})
-      # Special case gmail.com as it is potentially a popular OpenID
-      # provider and, unlike yahoo.com, where it works automatically, Google
-      # have hidden their OpenID endpoint somewhere obscure this making it
-      # somewhat less user friendly.
-      "https://www.google.com/accounts/o8/id"
-    else
-      openid_url
-    end
-  end
-
-  ##
-  # process a successful login
-  def successful_login(user, referer = nil)
-    session[:user] = user.id
-    session[:fingerprint] = user.fingerprint
-    session_expires_after 28.days if session[:remember_me]
-
-    target = referer || session[:referer] || url_for(:controller => :site, :action => :index)
-
-    # The user is logged in, so decide where to send them:
-    #
-    # - If they haven't seen the contributor terms, send them there.
-    # - If they have a block on them, show them that.
-    # - If they were referred to the login, send them back there.
-    # - Otherwise, send them to the home page.
-    if !user.terms_seen
-      redirect_to :action => :terms, :referer => target
-    elsif user.blocked_on_view
-      redirect_to user.blocked_on_view, :referer => target
-    else
-      redirect_to target
-    end
-
-    session.delete(:remember_me)
-    session.delete(:referer)
-  end
-
-  ##
-  # process a failed login
-  def failed_login(message, username = nil)
-    flash[:error] = message
-
-    redirect_to :action => "login", :referer => session[:referer],
-                :username => username, :remember_me => session[:remember_me]
-
-    session.delete(:remember_me)
-    session.delete(:referer)
-  end
-
-  ##
   #
   def unconfirmed_login(user)
     session[:token] = user.tokens.create.token
@@ -696,15 +579,6 @@ class UsersController < ApplicationController
     @user = User.find_by(:display_name => params[:display_name])
   rescue ActiveRecord::RecordNotFound
     redirect_to :action => "view", :display_name => params[:display_name] unless @user
-  end
-
-  ##
-  #
-  def disable_terms_redirect
-    # this is necessary otherwise going to the user terms page, when
-    # having not agreed already would cause an infinite redirect loop.
-    # it's .now so that this doesn't propagate to other pages.
-    flash.now[:skip_terms] = true
   end
 
   ##
