+++ app/views/sessions/destroy.html.erb
