+++ config/locales/en.yml
@@ -1586,6 +1586,59 @@ en:
       as_unread: "Message marked as unread"
     destroy:
       destroyed: "Message deleted"
+  sessions:
+    new:
+      title: "Login"
+      heading: "Login"
+      email or username: "Email Address or Username:"
+      password: "Password:"
+      openid_html: "%{logo} OpenID:"
+      remember: "Remember me"
+      lost password link: "Lost your password?"
+      login_button: "Login"
+      register now: Register now
+      with username: "Already have an OpenStreetMap account? Please login with your username and password:"
+      with external: "Alternatively, use a third party to login:"
+      new to osm: New to OpenStreetMap?
+      to make changes: To make changes to the OpenStreetMap data, you must have an account.
+      create account minute: Create an account. It only takes a minute.
+      no account: Don't have an account?
+      account not active: "Sorry, your account is not active yet.<br />Please use the link in the account confirmation email to activate your account, or <a href=\"%{reconfirm}\">request a new confirmation email</a>."
+      account is suspended: Sorry, your account has been suspended due to suspicious activity.<br />Please contact the <a href="%{webmaster}">webmaster</a> if you wish to discuss this.
+      auth failure: "Sorry, could not log in with those details."
+      openid_logo_alt: "Log in with an OpenID"
+      auth_providers:
+        openid:
+          title: Login with OpenID
+          alt: Login with an OpenID URL
+        google:
+          title: Login with Google
+          alt: Login with a Google OpenID
+        facebook:
+          title: Login with Facebook
+          alt: Login with a Facebook Account
+        windowslive:
+          title: Login with Windows Live
+          alt: Login with a Windows Live Account
+        github:
+          title: Login with GitHub
+          alt: Login with a GitHub Account
+        wikipedia:
+          title: Login with Wikipedia
+          alt: Login with a Wikipedia Account
+        yahoo:
+          title: Login with Yahoo
+          alt: Login with a Yahoo OpenID
+        wordpress:
+          title: Login with Wordpress
+          alt: Login with a Wordpress OpenID
+        aol:
+          title: Login with AOL
+          alt: Login with an AOL OpenID
+    destroy:
+      title: "Logout"
+      heading: "Logout from OpenStreetMap"
+      logout_button: "Logout"
   shared:
     markdown_help:
       title_html: Parsed with <a href="https://kramdown.gettalong.org/quickref.html">kramdown</a>
@@ -2221,58 +2274,6 @@ en:
     destroy:
       flash: "Destroyed the client application registration"
   users:
-    login:
-      title: "Login"
-      heading: "Login"
-      email or username: "Email Address or Username:"
-      password: "Password:"
-      openid_html: "%{logo} OpenID:"
-      remember: "Remember me"
-      lost password link: "Lost your password?"
-      login_button: "Login"
-      register now: Register now
-      with username: "Already have an OpenStreetMap account? Please login with your username and password:"
-      with external: "Alternatively, use a third party to login:"
-      new to osm: New to OpenStreetMap?
-      to make changes: To make changes to the OpenStreetMap data, you must have an account.
-      create account minute: Create an account. It only takes a minute.
-      no account: Don't have an account?
-      account not active: "Sorry, your account is not active yet.<br />Please use the link in the account confirmation email to activate your account, or <a href=\"%{reconfirm}\">request a new confirmation email</a>."
-      account is suspended: Sorry, your account has been suspended due to suspicious activity.<br />Please contact the <a href="%{webmaster}">webmaster</a> if you wish to discuss this.
-      auth failure: "Sorry, could not log in with those details."
-      openid_logo_alt: "Log in with an OpenID"
-      auth_providers:
-        openid:
-          title: Login with OpenID
-          alt: Login with an OpenID URL
-        google:
-          title: Login with Google
-          alt: Login with a Google OpenID
-        facebook:
-          title: Login with Facebook
-          alt: Login with a Facebook Account
-        windowslive:
-          title: Login with Windows Live
-          alt: Login with a Windows Live Account
-        github:
-          title: Login with GitHub
-          alt: Login with a GitHub Account
-        wikipedia:
-          title: Login with Wikipedia
-          alt: Login with a Wikipedia Account
-        yahoo:
-          title: Login with Yahoo
-          alt: Login with a Yahoo OpenID
-        wordpress:
-          title: Login with Wordpress
-          alt: Login with a Wordpress OpenID
-        aol:
-          title: Login with AOL
-          alt: Login with an AOL OpenID
-    logout:
-      title: "Logout"
-      heading: "Logout from OpenStreetMap"
-      logout_button: "Logout"
     lost_password:
       title: "Lost password"
       heading: "Forgotten Password?"
